add in custom hook and start updating examples
